it seems that rendering once on a wheel event is not enough..
in case you change the zoom direction and you are zooming very slow, you
will notice one zoom-step into the wrong direction. "Fixed" it by activating
animationloop and deactivating it immediately.

I have the feeling that the issue is the symptom of another problem (probably
wrong instruction order) which we haven't discovered, yet